Project: upgrade all packages in project
I decided there is no time better than now to do the upgrade. Webpack has not been upgraded to 4.X so we won't get heckled about vulnerabilities. There is still a good amount of packages cleanup that needs to be done, but for now it feels good to be current with all.