Database: complete sample unit database migrations
I went with using the sample unit id as the actual id for the record. I am hoping that there is a list of all of them that starts from one. If not this can be changed later, but for now it can be tested with our small sample of data. I made sure to add keys to enforce data structure.